fixed test issue (moved state stuff)
