+++ app/src/util/util.ts
@@ -1,7 +1,5 @@
 import { ColumnInfo, RGB, assertEventSettings, assertPointSettings } from "@tsconline/shared";
 import Color from "color";
-import { useContext } from "react";
-import { context } from "../state";
 
 export function discardTscPrefix(name: string | null) {
   if (!name) {
@@ -32,45 +30,6 @@ export function checkIfDataIsInRange(minDataAge: number, maxDataAge: number, use
     return true;
   }
   return (minDataAge > userTopAge && minDataAge < userBaseAge) || (maxDataAge < userBaseAge && maxDataAge > userTopAge);
-}
-
-export function checkIfDccDataIsInRange(dccColumn: ColumnInfo, userTopAge: number, userBaseAge: number) {
-  const { state } = useContext(context);
-  if (userBaseAge <= userTopAge) {
-    return false;
-  }
-  let reachedFirstRef = false;
-  while (!reachedFirstRef) {
-    if (
-      !(
-        (dccColumn.minAge <= userTopAge && dccColumn.maxAge >= userBaseAge) ||
-        (dccColumn.minAge > userTopAge && dccColumn.minAge < userBaseAge) ||
-        (dccColumn.maxAge < userBaseAge && dccColumn.maxAge > userTopAge)
-      )
-    ) {
-      return false;
-    }
-    if (dccColumn.columnDisplayType === "Event") {
-      assertEventSettings(dccColumn.columnSpecificSettings);
-    } else if (dccColumn.columnDisplayType === "Point") {
-      assertPointSettings(dccColumn.columnSpecificSettings);
-    } else {
-      console.warn("WARNING: dccColumn is not a valid column type");
-      return false;
-    }
-    //reached end of ref list
-    if (!dccColumn.columnSpecificSettings.dualColCompColumnRef) {
-      reachedFirstRef = true;
-      break;
-    }
-    const refCol = state.settingsTabs.columnHashMap.get(dccColumn.columnSpecificSettings.dualColCompColumnRef);
-    if (!refCol) {
-      console.log("WARNING: tried to get reference while checking dcc column, but is undefined");
-      return false;
-    }
-    dccColumn = refCol;
-  }
-  return true;
 }
 
 export function checkIfDccColumn(column: ColumnInfo) {
